ref #247 exceptions for delete_challenge method added

+++ app/javascript/dao/challenge_dao.py
@@ -1,14 +1,13 @@
-from app.javascript.exceptions.dao_get_challenge_exception import dao_get_challenge_exception
+from app.javascript.exceptions.challenge_dao_exception import challenge_dao_exception
 from ..models_js import JavascriptChallenge, db
 
 
 def get_challenge(id):
-    challenge = JavascriptChallenge.query.filter_by(id==JavascriptChallenge.id).first()
+    challenge = JavascriptChallenge.query.filter(id == JavascriptChallenge.id).first()
     if challenge == None: 
-        raise dao_get_challenge_exception(f'the id does not exist or is null',dao_get_challenge_exception.HTTP_NOT_FOUND)  
+        raise challenge_dao_exception(f'the id does not exist or is null',challenge_dao_exception.HTTP_NOT_FOUND)  
     return challenge
     
-
 def get_all_challenges():
     challenges = JavascriptChallenge.query.all()
     return challenges
@@ -20,7 +19,9 @@ def save_challenge(code, test_code, repair_objective,complexity,best_score):
     return challenge
 
 def delete_challenge(id):
-    challenge = JavascriptChallenge.query.filter_by(id==JavascriptChallenge.id).first()
+    challenge = JavascriptChallenge.query.filter(id==JavascriptChallenge.id).first()
+    if challenge == None:
+        raise challenge_dao_exception(f'the id does not exist or is null',challenge_dao_exception.HTTP_NOT_FOUND)
     db.session.delete(challenge)
     db.session.commit()
     return challenge
